+++ ui/app/actions.js
@@ -33,11 +33,7 @@ var actions = {
   ALERT_CLOSE: 'UI_ALERT_CLOSE',
   showAlert: showAlert,
   hideAlert: hideAlert,
-  QR_SCANNER_OPEN: 'UI_QR_SCANNER_OPEN',
-  QR_SCANNER_CLOSE: 'UI_QR_SCANNER_CLOSE',
   QR_CODE_DETECTED: 'UI_QR_CODE_DETECTED',
-  showQrScanner,
-  hideQrScanner,
   qrCodeDetected,
   // network dropdown open
   NETWORK_DROPDOWN_OPEN: 'UI_NETWORK_DROPDOWN_OPEN',
@@ -310,7 +306,6 @@ var actions = {
   CLEAR_PENDING_TOKENS: 'CLEAR_PENDING_TOKENS',
   setPendingTokens,
   clearPendingTokens,
-  scanQrCode,
 }
 
 module.exports = actions
@@ -1813,25 +1808,12 @@ function hideAlert () {
   }
 }
 
-function showQrScanner () {
-  return {
-    type: actions.QR_SCANNER_OPEN,
-  }
-}
-
 function qrCodeDetected (qrCodeData) {
   return {
     type: actions.QR_CODE_DETECTED,
     value: qrCodeData,
   }
 }
-
-function hideQrScanner () {
-  return {
-    type: actions.QR_SCANNER_CLOSE,
-  }
-}
-
 
 function showLoadingIndication (message) {
   return {
@@ -2273,11 +2255,5 @@ function setPendingTokens (pendingTokens) {
 function clearPendingTokens () {
   return {
     type: actions.CLEAR_PENDING_TOKENS,
-  }
-}
-
-function scanQrCode () {
-  return (dispatch, getState) => {
-    dispatch(actions.showQrScanner())
   }
 }
