+++ ui/app/app.js
@@ -38,9 +38,6 @@ const AccountMenu = require('./components/account-menu')
 const Modal = require('./components/modals/index').Modal
 // Global Alert
 const Alert = require('./components/alert')
-
-const QrScanner = require('./components/qr-scanner')
-
 const AppHeader = require('./components/app-header')
 
 import UnlockPage from './components/pages/unlock-page'
@@ -133,8 +130,6 @@ class App extends Component {
 
         // global alert
         h(Alert, {visible: this.props.alertOpen, msg: alertMessage}),
-
-        h(QrScanner, {visible: this.props.qrScannerOpen}),
 
         h(AppHeader),
 
@@ -274,7 +269,6 @@ App.propTypes = {
   currentView: PropTypes.object,
   sidebarOpen: PropTypes.bool,
   alertOpen: PropTypes.bool,
-  qrScannerOpen: PropTypes.bool,
   hideSidebar: PropTypes.func,
   isMascara: PropTypes.bool,
   isOnboarding: PropTypes.bool,
@@ -311,7 +305,6 @@ function mapStateToProps (state) {
     networkDropdownOpen,
     sidebarOpen,
     alertOpen,
-    qrScannerOpen,
     alertMessage,
     isLoading,
     loadingMessage,
@@ -339,7 +332,6 @@ function mapStateToProps (state) {
     networkDropdownOpen,
     sidebarOpen,
     alertOpen,
-    qrScannerOpen,
     alertMessage,
     isLoading,
     loadingMessage,
@@ -370,7 +362,6 @@ function mapStateToProps (state) {
     isMouseUser: state.appState.isMouseUser,
     betaUI: state.metamask.featureFlags.betaUI,
     isRevealingSeedWords: state.metamask.isRevealingSeedWords,
-    Qr: state.appState.Qr,
     welcomeScreenSeen: state.metamask.welcomeScreenSeen,
 
     // state needed to get account dropdown temporarily rendering from app bar
