+++ ui/app/components/modals/qr-scanner/index.js
@@ -1,0 +1,2 @@
+import QrScanner from './qr-scanner.container'
+module.exports = QrScanner
