--- ui/app/components/qr-scanner/index.js
@@ -1,2 +1,0 @@
-import QrScanner from './qr-scanner.component'
-module.exports = QrScanner
