+++ ui/app/components/send/send.container.js
@@ -28,7 +28,7 @@ import {
   updateSendTokenBalance,
   updateGasData,
   setGasTotal,
-  scanQrCode,
+  showModal,
 } from '../../actions'
 import {
   resetSendState,
@@ -93,7 +93,7 @@ function mapDispatchToProps (dispatch) {
     },
     updateSendErrors: newError => dispatch(updateSendErrors(newError)),
     resetSendState: () => dispatch(resetSendState()),
-    scanQrCode: () => dispatch(scanQrCode()),
+    scanQrCode: () => dispatch(showModal({ name: 'QR_SCANNER' })),
     updateSendTo: (to, nickname) => dispatch(updateSendTo(to, nickname)),
   }
 }
